fix mistake and add functions in SmartSampling
